ml-service/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6
sse-starlette==1.6.5
pydantic==2.4.2
aiohttp==3.8.5
python-dotenv==1.0.0
